Update paper_review.md with new plan
Before, the system for paper review didn't account for weeks with only one paper-review day. As a result, it was ambiguous as to what type of review should be done for those days.

This proposed change would instead have the alternating between critical and comprehension reviews happen for each _day_ of papers, rather than each week. As a result, if you did a critical review for paper 4a, you would do a comprehension review for either 5a or 5b.